requirement.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.4.0
astroid==2.9.0
atomicwrites==1.4.0
attrs==21.3.0
autopep8==1.6.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==4.4.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.2
dj-rest-auth==2.2.1
Django==3.1.7
django-allauth==0.45.0
django-celery-results==2.2.0
django-cors-headers==3.8.0
django-filter==2.4.0
django-froala-editor==3.2.2
django-rest-auth==0.9.5
django-restframework==0.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==11.0.0
filelock==3.0.12
idna==3.2
inflection==0.3.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.11
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
pep8==1.7.1
Pillow==8.3.1
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.1.0
pylint==2.12.1
pyparsing==3.0.6
pytest==6.2.5
pytest-assert-utils==0.3.0
pytest-common-subject==1.0.5
pytest-drf==1.1.2
pytest-fixture-order==0.1.3
pytest-lambda==1.2.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.18
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==1.3.0
virtualenv==20.4.7
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
wrapt==1.13.3
